refactor UserPaymentRequest
